+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/Constants.kt
@@ -36,6 +36,9 @@ object Constants {
     }
 
     object Constraints {
+        const val CLOSEST_VERSION = "closest"
+        const val LATEST_VERSION = "latest"
+        const val PLATFORM_VERSION = "platform"
         val MINIMAL_GRADLE_VERSION: GradleVersion = GradleVersion.version("8.2")
         val MINIMAL_INTELLIJ_PLATFORM_BUILD_NUMBER = "223".toVersion()
         val MINIMAL_INTELLIJ_PLATFORM_VERSION = "2022.3".toVersion()