+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/models/Coordinates.kt
@@ -2,7 +2,17 @@
 
 package org.jetbrains.intellij.platform.gradle.models
 
+import org.jetbrains.intellij.platform.gradle.Constants.Locations
+import java.net.URL
+
 data class Coordinates(val groupId: String, val artifactId: String) {
 
     override fun toString() = "$groupId:$artifactId"
+}
+
+internal fun Coordinates.resolveLatestVersion(repositoryUrl: String = Locations.MAVEN_REPOSITORY): String? {
+    val host = repositoryUrl.trimEnd('/')
+    val path = toString().replace(':', '/').replace('.', '/')
+    val url = URL("$host/$path/maven-metadata.xml")
+    return decode<MavenMetadata>(url).versioning?.latest
 }