+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -369,6 +369,7 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
                 is SplitModeAware,
                 is TestableAware,
                     -> true
+
                 else -> false
             }
         }.configureEach(project::preconfigureTask)