+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - Rename `jetBrainsCdn()` repository helper to `jetbrainsIdeInstallers()`
 - Rename `binaryReleasesAndroidStudio()` repository helper to `androidStudioInstallers()`
+- Rewrite the latest/closest version resolution mechanism for performance reasons
+- Rewrite the latest Gradle plugin check for performance reasons
 
 ### Fixed